More project organisation - put sparse and WLS into their own files
